adicionado estrutura basica dos sockets
